consolide les résultats
