+++ documentation/backlog.md
@@ -17,6 +17,9 @@
 notes : Pour solder les comptes du bilan, les comptes débiteurs sont crédités et les compte créditeurs sont débités. comme au bilan l'actif est égales au passif, on peut donc passer une écriture unique ou passer deux écritures en utilisant le compte 891
 
 ## US8 Consolidation
+- [ ] affectation resultat precedent 101
+- [ ] reprendre les textes d'aides
+- [ ] impot de l'exercice en charge constaté ?
 - [ ] AN- en une écriture - (P172 la compta pour les nuls) - avec vérif. équilibre
   - [ ] affectation réserve libre 106
 - [ ] passer les comptes 2 et 3 ?
