+++ plan-comptable.js
@@ -6,11 +6,12 @@ export const PLAN_COMPTABLE = {
     "275000": ["dépôts et cautionnements versés", "cautions"],
     "370000": ["stocks de marchandises", "inventaire"],
     "404000": ["fournisseurs d'immobilisations"],
+    "444000": ["état - impôts sur les bénéfices"],
     "467000": ["autres comptes débiteurs ou créditeurs", "remboursements", "prêts"],
     "512000": ["banques"],
     "530000": ["caisse"],
     "602600": ["emballages"],
-    "603000": ["variations des stocks"],
+    "603700": ["variation des stocks de marchandises"],
     "604000": ["achats d'études et prestations de services", "achats prestations"],
     "606000": ["achats non stockés de matière et fournitures", "fournitures", "décorations", "énergies"],
     "607000": ["achats de marchandises", "marchandises"],
@@ -24,6 +25,7 @@ export const PLAN_COMPTABLE = {
     "625000": ["déplacements, missions et réceptions", "déplacements", "restauration", "hébergements"],
     "626000": ["frais postaux et de télécommunications", "internet", "frais postaux", "télécommunications", "domiciliations"],
     "627000": ["services bancaires et assimilés", "services bancaires"],
+    "695000": ["impôts sur les bénéfices"],
     "706000": ["prestations de services", "formations"],
     "707000": ["ventes de marchandises", "ventes"],
     "754100": ["dons manuels", "dons"],
